+++ packages/web/src/package/ui/KeyDialog/editor/PluralEditor.tsx
@@ -49,7 +49,11 @@ export const PluralEditor = ({
       render={({ content, variant, exampleValue }) => {
         const variantOrOther = variant || 'other';
         return (
-          <EditorWrapper data-cy="translation-editor" data-cy-variant={variant}>
+          <EditorWrapper
+            data-cy="translation-editor"
+            data-cy-variant={variant}
+            data-cy-language={locale}
+          >
             <Editor
               mode={editorMode}
               value={content}